quarkus: no main method anymore, using CDI's @Observes StartupEvent instead; removing some static fields, because it causes issues with ahead-of-time compilation in hte native image
